Uploading dpy-camera plugin for sketch up
Uploading dpy-camera plugin for sketch up - a tool for navigating in
sketch up as if in computer shooter game, using arrow keys to move
camera and the mouse to rotate it. There is extra functionality called
GunMode and when it's turned On you can shoot at and kill (erase)
objects from sketch up model. :) Enjoy!